LayoutTests/platform/mac/editing/caret/caret-position-sideways-lr-expected.txt

This is a test of writingsideways in vertical text
with multiline text and alongwordthat'slong andanotherword.
Bug 286961 : 62,45,20,2
Click at 5, 5 for left above first line : 17,22,25,2
Click at 5, 275 for right above first line : 17,22,25,2
Click at 25, 5 for left of first line : 17,22,25,2
Click at 25, 275 for right of first line : 17,238,25,2
Click at 45, 5 for left of second line : 42,22,20,2
Click at 45, 275 for right of second line : 42,249,20,2
Click at 65, 5 for left of third line : 62,22,20,2
Click at 65, 275 for right of third line : 62,177,20,2
Click at 85, 5 for left of fourth line : 82,22,20,2
Click at 85, 275 for right of fourth line : 82,263,20,2
Click at 105, 5 for left of fifth line : 102,22,19,2
Click at 105, 275 for right of fifth line : 102,69,19,2
Click at 125, 5 for left of sixth line : 121,22,20,2
Click at 125, 275 for right of sixth line : 121,261,20,2
Click at 145, 5 for left of seventh line : 141,22,20,2
Click at 145, 275 for right of seventh line : 141,201,20,2
Click at 170, 5 for left below last line : 141,201,20,2
Click at 170, 275 for right below last line : 141,201,20,2
Click at 25, 30 for inside first root inline fragment : 17,33,25,2
Click at 125, 30 for inside middle fragment of root inline fragment : 121,33,20,2
Click at 45, 30 for inside middle fragment of non-root inline fragment : 42,33,20,2
Click at 65, 30 for inside last fragment of non-root inline fragment : 62,33,20,2
Click at 85, 220 for inside unfragmented non-root inline : 82,226,20,2
